+++ app/User.php
@@ -11,6 +11,8 @@ class User extends Authenticatable
     use Notifiable;
     protected $table = 'Users';
     protected $primaryKey = 'email';
+    public $incrementing = false;
+
     /**
      * The attributes that are mass assignable.
      *
