+++ resources/views/ForgotPassword/emailSent.blade.php
@@ -1,4 +1,5 @@
-@extends('layout')
+@extends('Layout/layout')
+@include('Layout/header')
 @section('content')
 <h1>Password Reset Link Sent</h1>
 <a href="{{ route('inventory') }}"><button type="button">Back to Inventory</button></a>
