+++ resources/views/Layout/layout.blade.php
