+++ routes/web.php
@@ -19,8 +19,15 @@ Route::get('/', function () {
 Route::group(['middleware' => 'web'], function () {
     Route::get('/', 'InventoryController@inventoryPage')->name('inventory');
     Route::post('/getItems/{index?}', 'InventoryController@getItems')->name('getItems');
-    Route::post('/addToCart', 'InventoryController@addToCart')->name('addToCart');
-    Route::get('/cart', 'InventoryController@cart')->name('cart');
+    
+    Route::post('/addToCart', 'CartController@addToCart')->name('addToCart');
+    Route::get('/cart', 'CartController@cartPage')->name('cart');
+    
+    Route::get('/forgotPassword', 'ForgotPasswordController@forgotPassword')->name('forgotPassword');
+    Route::post('/forgotPassword', 'ForgotPasswordController@submitForgotPassword')->name('submitForgotPassword');
+    
+    Route::get('/resetPassword/{token}', 'ResetPasswordController@resetPassword')->name('resetPassword');
+    Route::post('/resetPassword/{token}', 'ResetPasswordController@submitResetPassword')->name('submitResetPassword');
 
     Auth::routes();
 
@@ -30,10 +37,4 @@ Route::group(['middleware' => 'web'], function () {
     
     Route::get('/register', 'RegisterController@register')->name('register');
     Route::post('/register', 'RegisterController@submitRegister')->name('submitRegister');
-    
-    Route::get('/forgotPassword', 'ForgotPasswordController@forgotPassword')->name('forgotPassword');
-    Route::post('/forgotPassword', 'ForgotPasswordController@submitForgotPassword')->name('submitForgotPassword');
-    
-    Route::get('/resetPassword/{token}', 'ResetPasswordController@resetPassword')->name('resetPassword');
-    Route::post('/resetPassword/{token}', 'ResetPasswordController@submitResetPassword')->name('submitResetPassword');
 });
